Update README.md
The TriageCase.py file contains a simulation for a hospital which employs adaptable priority queue. The Age is used to aasign priority to patients arriving in the hospital

The CandyDispenser.py is simulation of candy dispenser which employes stack data structure in LIFO mechanism to dispense candies or add candies to the candy stack.

The dieroll.py is asimulation model for a die rolled a thousand times. The experiment simulates the probability of playing dice. The faces are randomly assigned to random occurences in the thousand rolls, frquency taken and then percentage of each face frequency computed.

The automata.py deals with theory of computation and it simulates a Deterministic Finite Automata which takes some alphabet, then creates two Languages from the alphabet and then evaluates strings under certain rules to acertain whether those strings meet the criteria of being in a language. strings are either accepted or rejected.